18/pyproject.toml
[tool.poetry]
name = "18"
version = "0.1.0"
description = ""
authors = ["Marek Orłowski <16357457+ormarek@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
"colorgram.py" = "^1.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
